refactor: rename notificationContent module to notificationCard
